move requestTransformer to utils
